Rename markdown to blog
