+++ sites/svelte.dev/src/routes/blog/+page.server.js
@@ -1,9 +1,9 @@
-import { get_index } from '$lib/server/markdown';
+import { get_index } from '$lib/server/blog';
 
 export const prerender = true;
 
 export async function load() {
 	return {
-		posts: get_index()
+		posts: get_index(),
 	};
 }
