+++ sites/svelte.dev/tsconfig.json
@@ -1,3 +1,8 @@
 {
 	"extends": "./.svelte-kit/tsconfig.json",
+	"compilerOptions": {
+		"allowJs": true,
+		"checkJs": true,
+		"allowSyntheticDefaultImports": true
+	}
 }
